afronectar/pyproject.toml
[tool.poetry]
name = "afronectar"
version = "0.1.0"
description = "app for testing purpose for nginx reverse proxy"
authors = ["abdiwali maxamed"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
gunicorn = "^22.0.0"
django-mptt = "^0.16.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
